Implement Phase 17: Cross-service spec compliance audit fixes
Comprehensive audit of all 7 specs against their implementations found
and fixed key gaps:

- LLM Agent: Added AuditLogger (was completely missing), audit endpoints,
  input validation for chat, fixed LLM_MODE env var name
- Expense API: Added DEMO_MODE guard on unprotected endpoint, fixed
  error message wording to match spec
- VC Issuer: Changed approvalLimit validation to require integer
- VC Wallet: Fixed holder binding to reject missing credentialSubject.id,
  filtered VerifiableCredential base type from error responses

All 167 tests passing, typecheck clean, 0 lint errors.

+++ src/vc-issuer/index.ts
@@ -46,7 +46,7 @@ const EmployeeCredentialRequestSchema = z.object({
 
 const FinanceApproverCredentialRequestSchema = z.object({
   subjectDid: z.string().startsWith('did:'),
-  approvalLimit: z.number().positive(),
+  approvalLimit: z.number().int().positive(),
   department: z.string().min(1),
 });
